refactor: extract dxf mid-layer helpers
